chore(openapisampler): tests splitting
closes #2
